Create Data Structures & Algorithms - Overview
Data Structure is a systematic way to organize data in order to use it efficiently.
Data Structure is a systematic way to organize data in order to use it efficiently. 

Characteristics of a Data Structure 
Correctness − Data structure implementation should implement its interface correctly.

Time Complexity − The running time or the execution time of operations of the data structure must be as small as possible.

Space Complexity − Memory usage of a data structure operation should be as little as possible.